Code for Hangman Game
Hangman is an old school favorite, a word game where the goal is simply to find the missing word. You will be presented with a number of blank spaces representing the missing letters you need to find.
Use the keyboard to guess a letter. If your chosen letter exists in the answer, then all places in the answer where that letter appear will be revealed. After you've revealed several letters, you may be able to guess what the answer is and fill in the remaining letters.
Be warned, every time you guess a letter wrong you loose a life and the hangman begins to appear, piece by piece. Solve the puzzle before the hangman dies.